requirements.txt
# ===============================
# Data processing
# ===============================
pandas==2.2.0
numpy==1.26.4

# ===============================
# Airflow
# ===============================
apache-airflow==2.8.0
apache-airflow-providers-apache-spark

# ===============================
# Spark
# ===============================
pyspark==3.5.1
delta-spark==3.1.0
# ===============================
# Database
# ===============================
sqlalchemy==1.4.52
psycopg2-binary==2.9.9

# ===============================
# File formats
# ===============================
pyarrow==15.0.0
openpyxl==3.1.2

# ===============================
# Utilities
# ===============================
python-dotenv==1.0.1
